moved linter command to its own handler file, linter switching methods  to their own file, and shifted over some npm helpers to lint-worker (for when we want them for react-codemirror). Also set vscode messages to be constants
