+++ packages/lint-worker/src/helpers.ts
@@ -4,6 +4,7 @@ import {
   Neo4jFunction,
   Neo4jProcedure,
 } from '@neo4j-cypher/language-support';
+import axios from 'axios';
 import { DbSchema as DbSchemaV1 } from 'languageSupport-next.13';
 
 const oldLinter = '5.20.0';
@@ -50,4 +51,37 @@ export function serverVersionToLinter(serverVersion: string) {
 export function linterFileToServerVersion(fileName: string) {
   const linterFileRegex = /^([\d.]+)-lintWorker-([\d.]+)\.cjs$/;
   return fileName ? fileName.match(linterFileRegex)?.[1] : undefined;
+}
+
+//The data object we get from npm contains more fields than this, but we only need dist-tags here
+export type NpmData = {
+  'dist-tags'?: Record<string, string>;
+};
+
+export type NpmRelease = {
+  tag: string;
+  version: string;
+};
+
+export const npmTagToLinterVersion = (tag: string) =>
+  tag.match(/^neo4j-([\d.]+)$/)?.[1];
+
+export async function getTaggedRegistryVersions(): Promise<NpmRelease[]> {
+  const registryUrl = 'https://registry.npmjs.org/@neo4j-cypher/lint-worker';
+  try {
+    const response = await axios.get<NpmData>(registryUrl);
+    const data: NpmData = response.data;
+    const taggedVersions: { tag: string; version: string }[] = [];
+    if (data !== null && data['dist-tags'] !== null) {
+      for (const [tag, version] of Object.entries(data['dist-tags'])) {
+        if (typeof tag === 'string' && typeof version === 'string') {
+          taggedVersions.push({ tag, version });
+        }
+      }
+    }
+
+    return taggedVersions;
+  } catch (error) {
+    return [];
+  }
 }
