+++ packages/vscode-extension/src/constants.ts
@@ -24,6 +24,9 @@ export const CONSTANTS = {
     },
   },
   MESSAGES: {
+    GLOBALSTORAGE_READ_FAILED: 'Failed to read neo4j globalStorage directory.',
+    LINTER_DOWNLOAD_FAILED:
+      'Linter download failed, reverting to best match from currently downloaded linter versions.',
     CONNECTED_MESSAGE: 'Connected to Neo4j.',
     DISCONNECTED_MESSAGE: 'Disconnected from Neo4j.',
     RECONNECTED_MESSAGE: 'Reconnected to Neo4j.',
